Hide unused padding (#133)
With this snippet the icon and label are never shown at the same time, but the padding between them still exists between them. Also tack on an !important to the bit that hides the icons. I could not get it to work without it which is just bizarre because it really should.